api-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.12.0
pydantic-settings==2.1.0
google-genai>=1.0.0
boto3==1.34.0
numpy>=1.24.0,<2
pdf2image==1.17.0
Pillow==10.1.0
opencv-python==4.8.1.78
